+++ test/index.js
@@ -1,0 +1,4 @@
+[
+  "./unit/index",
+  "./integration/index",
+].forEach(require);