+++ test/unit/index.js
@@ -1,0 +1,4 @@
+[
+  "./test",
+  "./test-deprecated-export",
+].forEach(require);